requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils==0.41.1
pymysql
pytest
pytest-mock
httpx
cryptography
stripe
python-dotenv
pyjwt